Add `show_nn` function and refactor print statements to return
This commit introduces a new `show_nn` function to display the neural network. In addition, all the print statements have been refactored to return statements which will be more beneficial debugging and testing. Lastly, relevant variable names were changed to improve readability and understanding of the code. All changes aim to improve the overall robustness and efficiency of the program.