Refactor help command and add output list in main function
Introduced an output list in the main function of the neuron network module to keep track of the generated outputs during each epoch. This will aid in debugging and performance analysis.
In the neural network interface, the help command which shows all available commands to the user was refactored. It previously existed as a dictionary in the commands.py file but is now in its own dedicated file cmd_help.py making it cleaner and easier to maintain. Functions in commands.py file was also simplified to use this new structure. I also changed the help command name to get_help to avoid Python built-in function name conflict.