Add new features to Discord-bot and NeuronalNetwork and fix bugs
This commit includes adding new features to the Beatrix discord bot that allows it to handle user commands and reply to user messages. Also, the multilayer NeuronalNetwork has been expanded and updated. New features include the addition of tkinter for creating graphical user interface and commands that enable interaction with a database (SQLite3). Plus, unused code and bugs were cleaned and fixed hence increasing the efficiency of the code. Addition of new files in NeuronalNetwork ensures software meets the requirements of creating a neuronal network, providing a user interface and a connection to a database.